requirements.txt
PyJWT==2.6.0
cryptography==38.0.4
cffi==1.15.1
requests==2.28.1